requirements.txt
Package                     Version
--------------------------- -----------
absl-py                     0.13.0
aiohttp                     3.7.4.post0
alembic                     1.4.1
appdirs                     1.4.4
async-timeout               3.0.1
atpublic                    2.3
attrs                       21.2.0
cached-property             1.5.2
cachetools                  4.2.2
certifi                     2021.5.30
cffi                        1.14.5
chardet                     4.0.0
click                       8.0.1
cloudpickle                 1.6.0
colorama                    0.4.4
commonmark                  0.9.1
configobj                   5.0.6
databricks-cli              0.14.3
dataclasses                 0.7
decorator                   4.4.2
dictdiffer                  0.8.1
diskcache                   5.2.1
distro                      1.5.0
docker                      5.0.0
dpath                       2.0.1
dulwich                     0.20.23
dvc                         2.4.3
efficientnet-pytorch        0.6.3
entrypoints                 0.3
Flask                       2.0.1
flatten-dict                0.4.0
flufl.lock                  3.2
fsspec                      2021.6.1
ftfy                        6.0.3
funcy                       1.16
future                      0.18.2
gitdb                       4.0.7
GitPython                   3.1.18
google-auth                 1.32.0
google-auth-oauthlib        0.4.4
grandalf                    0.6
greenlet                    1.1.0
grpcio                      1.38.1
gunicorn                    20.1.0
idna                        2.10
idna-ssl                    1.1.0
importlib-metadata          4.6.0
itsdangerous                2.0.1
Jinja2                      3.0.1
jsonpath-ng                 1.5.2
mailchecker                 4.0.9
Mako                        1.1.4
Markdown                    3.3.4
MarkupSafe                  2.0.1
mlflow                      1.18.0
multidict                   5.1.0
munch                       2.5.0
nanotime                    0.5.2
networkx                    2.5.1
numpy                       1.19.5
oauthlib                    3.1.1
packaging                   20.9
pandas                      1.1.5
pathspec                    0.8.1
phonenumbers                8.12.26
Pillow                      8.2.0
pip                         21.1.3
ply                         3.11
pretrainedmodels            0.7.4
prometheus-client           0.11.0
prometheus-flask-exporter   0.18.2
protobuf                    3.17.3
psutil                      5.8.0
pyasn1                      0.4.8
pyasn1-modules              0.2.8
pycparser                   2.20
pyDeprecate                 0.3.0
pydot                       1.4.2
pygit2                      1.6.1
Pygments                    2.9.0
pygtrie                     2.4.2
pyparsing                   2.4.7
python-benedict             0.24.0
python-dateutil             2.8.1
python-editor               1.0.4
python-fsutil               0.5.0
python-slugify              5.0.2
pytorch-lightning           1.3.7.post0
pytz                        2021.1
PyYAML                      5.4.1
querystring-parser          1.2.4
requests                    2.25.1
requests-oauthlib           1.3.0
rich                        10.4.0
rsa                         4.7.2
ruamel.yaml                 0.17.10
ruamel.yaml.clib            0.2.4
segmentation-models-pytorch 0.1.3
setuptools                  57.0.0
shortuuid                   1.0.1
shtab                       1.3.6
six                         1.16.0
smmap                       4.0.0
SQLAlchemy                  1.4.20
sqlparse                    0.4.1
tabulate                    0.8.9
tensorboard                 2.4.1
tensorboard-plugin-wit      1.8.0
text-unidecode              1.3
timm                        0.3.2
toml                        0.10.2
torch                       1.9.0
torchmetrics                0.4.0
torchvision                 0.10.0
tqdm                        4.61.1
typing-extensions           3.10.0.0
urllib3                     1.26.6
voluptuous                  0.12.1
wcwidth                     0.2.5
websocket-client            1.1.0
Werkzeug                    2.0.1
wheel                       0.36.2
xmltodict                   0.12.0
yarl                        1.6.3
zc.lockfile                 2.0
zipp                        3.4.1
